backend/requirements.txt
aiosqlite==0.22.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.124.4
greenlet==3.3.0
h11==0.16.0
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-jose==3.5.0
python-slugify==8.0.4
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
starlette==0.50.0
text-unidecode==1.3
typing_extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
